update
Aggiunto bootstrap su tutte le pagine
Ripulito index.php
Sistemati form secondo gli standard di bootstrap
Ripulite pagine d’avviso

+++ sitoEcommerce/rimuovi-prodotti.php
@@ -1,3 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<!-- Bootstrap -->
+<link rel="stylesheet" href="css/bootstrap.css">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>Rimuovi Prodotti</title>
+<body>
 <?php
 include('connessione_db.php');
 //Questa pagina viene utilizzata dall'amministratore per aggiungere o modificare prodotti dell'inventario
@@ -7,37 +14,51 @@ include('connessione_db.php');
 //La rimozione del prodotto si basa sulla ricerca per codice.
 //Se il codice inserito trova corrispondenza nella tabella prodotti utilizzo l'ID associato a quel codice prodotto
 //per rimuovere la riga dalla tabella
+if (!empty($_SESSION['email'])) {
+    if ($_SESSION['email'] != "ADMIN") {
+        stampaAvviso("Sezione non autorizzata", "index.php");
+    } else {
+        if (isset($_POST['rimuovi'])){
+          $codice = isset($_POST['codice']) ? clear($_POST['codice']) : false;
 
-if ($_SESSION['email'] != "ADMIN") {
-      echo 'Sezione non autorizzata.<br /><br /><a href="javascript:history.back();">Indietro</a>';
-} else {
-    if (isset($_POST['rimuovi'])){
-      $codice = isset($_POST['codice']) ? clear($_POST['codice']) : false;
 
+          if(empty($codice)) {
+            stampaAvviso("Riempi Campo Codice", "rimuovi-prodotti.php");
+        	} elseif (mysql_num_rows(mysql_query("SELECT * FROM prodotti WHERE codice='$codice'")) > 0) {
+            $id = mysql_result(mysql_query("SELECT id FROM prodotti WHERE codice='$codice'"), 0);
+            mysql_query("DELETE FROM prodotti WHERE id='$id'");
+            header("Location: index.php");
+          } else {
+            stampaAvviso("Prodotto non trovato in catalogo", "rimuovi-prodotti.php");
+          }
 
-      if(empty($codice)) {
-    		echo 'Riempi il campo Codice Prodotto.<br /><br /><a href="javascript:history.back();">Indietro</a>';
-    	} elseif (mysql_num_rows(mysql_query("SELECT * FROM prodotti WHERE codice='$codice'")) > 0) {
-        $id = mysql_result(mysql_query("SELECT id FROM prodotti WHERE codice='$codice'"), 0);
-        mysql_query("DELETE FROM prodotti WHERE id='$id'");
-        header("Location: index.php");
-      } else {
-        echo 'Prodotto non trovato in catalogo.<br /><br /><a href="javascript:history.back();">Indietro</a>';
-      }
-
-    }elseif (isset($_POST['home'])) {
-    header("location: index.php");
-    }else {
-      ?>
-      <div class="box">
-        <h2>Rimuovi Prodotto dal Catalogo </h2>
-        <form action="<?php echo $_SERVER['PHP_SELF']; ?>" method="POST">
-          <label><input type="text" class="label" name="codice" placeholder="Codice Prodotto"  maxlength="60" /></label><br />
-          <input type="submit" class="button" name="rimuovi" value="Rimuovi" />
-          <input type="submit" class="button" name="home" value="Torna alla Home" />
-        </form>
-      </div>
-      <?php
+        }elseif (isset($_POST['home'])) {
+        header("location: index.php");
+        }else {
+          ?>
+          <div clas="container">
+            <div class="box">
+              <h2>Rimuovi Prodotto dal Catalogo </h2><br/><br/>
+              <div class="form-group">
+              <form action="<?php echo $_SERVER['PHP_SELF']; ?>" method="POST">
+                <input type="text" class="form-control" name="codice" placeholder="Codice Prodotto"  maxlength="60" /><br/>
+              </div>
+              <div class="form-group" align=center>
+                <input type="submit" class="btn btn-default" name="rimuovi" value="Rimuovi" />
+                <input type="submit" class="btn btn-default" name="home" value="Torna alla Home" />
+            </div>
+            </form>
+          </div>
+        </div>
+          <?php
+        }
     }
-}
+} else {
+    header("location: index.php");
+  }
 ?>
+<!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
+<script src="js/jquery-1.11.3.min.js"></script>
+<script src="js/bootstrap.js"></script>
+</body>
+</html>
